Hacktoberfest 👀 (#153)
* Hacktoberfest entry

* Hacktoberfest entry

* Well.. uh.. this is embarrassing..

+++ runtime/internal/voice.js
@@ -11,7 +11,7 @@ var Logger = require('./logger.js').Logger
 var Config = require('../../config.json')
 
 exports.join = function (msg, suffix, bot) {
-  if (bot.VoiceConnections.length > 10) {
+  if (bot.VoiceConnections.length > config.settings.maxvcslots) {
     msg.channel.sendMessage('Sorry, all streaming slots are taken, try again later. :cry:')
   } else {
     var voiceCheck = bot.VoiceConnections.find((r) => r.voiceConnection.guild.id === msg.guild.id)
@@ -118,9 +118,11 @@ exports.join = function (msg, suffix, bot) {
         })
     } else {
       msg.reply('I am already streaming on this server in channel **' + voiceCheck.voiceConnection.channel.name + '**').then((m) => {
-        setTimeout(() => {
-          m.delete().catch((e) => Logger.error(e))
-        }, 3000)
+        if (config.autodeletemsg) {
+          setTimeout(() => {
+            m.delete().catch((e) => Logger.error(e))
+          }, config.deleteTimeout)
+        }
       })
     }
   }
@@ -187,9 +189,11 @@ function next (msg, suffix, bot) {
         connection.voiceConnection.getEncoder().setVolume(vol)
         encoder.once('end', () => {
           msg.channel.sendMessage('**' + list[msg.guild.id].info[0] + '** has ended!').then((m) => {
-            setTimeout(() => {
-              m.delete().catch((e) => Logger.error(e))
-            }, 3000)
+            if (config.autodeletemsg) {
+              setTimeout(() => {
+                m.delete().catch((e) => Logger.error(e))
+              }, config.deleteTimeout)
+            }
           })
           list[msg.guild.id].link.shift()
           list[msg.guild.id].info.shift()
@@ -198,16 +202,20 @@ function next (msg, suffix, bot) {
           list[msg.guild.id].skips.users = []
           if (list[msg.guild.id].link.length > 0) {
             msg.channel.sendMessage('Next up is **' + list[msg.guild.id].info[0] + '** requested by _' + list[msg.guild.id].requester[0] + '_').then((m) => {
-              setTimeout(() => {
-                m.delete().catch((e) => Logger.error(e))
-              }, 6000)
+              if (config.autodeletemsg) {
+                setTimeout(() => {
+                  m.delete().catch((e) => Logger.error(e))
+                }, config.deleteTimeoutLong)
+              }
             })
             next(msg, suffix, bot)
           } else {
             msg.channel.sendMessage('Playlist has ended, leaving voice.').then((m) => {
-              setTimeout(() => {
-                m.delete().catch((e) => Logger.error(e))
-              }, 3000)
+              if (config.autodeletemsg) {
+                setTimeout(() => {
+                  m.delete().catch((e) => Logger.error(e))
+                }, config.deleteTimeout)
+              }
             })
             connection.voiceConnection.disconnect()
           }
@@ -357,9 +365,11 @@ exports.request = function (msg, suffix, bot) {
     }, function (err, data) {
       if (err) {
         msg.channel.sendMessage('Something went wrong while requesting information about this playlist.').then((m) => {
-          setTimeout(() => {
-            m.delete().catch((e) => Logger.error(e))
-          }, 3000)
+          if (config.autodeletemsg) {
+            setTimeout(() => {
+              m.delete().catch((e) => Logger.error(e))
+            }, config.deleteTimeout)
+          }
         })
         Logger.error('Playlist failiure, ' + err)
         return
@@ -371,9 +381,11 @@ exports.request = function (msg, suffix, bot) {
   } else {
     fetch(suffix, msg).then((r) => {
       msg.channel.sendMessage(`Added **${r.title}** to the playlist.`).then((m) => {
-        setTimeout(() => {
-          m.delete().catch((e) => Logger.error(e))
-        }, 3000)
+        if (config.autodeletemsg) {
+          setTimeout(() => {
+            m.delete().catch((e) => Logger.error(e))
+          }, config.deleteTimeout)
+        }
       })
       if (r.autoplay === true) {
         next(msg, suffix, bot)
@@ -381,9 +393,11 @@ exports.request = function (msg, suffix, bot) {
     }).catch((e) => {
       Logger.error(e)
       msg.channel.sendMessage("I couldn't add that to the playlist.").then((m) => {
-        setTimeout(() => {
-          m.delete().catch((e) => Logger.error(e))
-        }, 3000)
+        if (config.autodeletemsg) {
+          setTimeout(() => {
+            m.delete().catch((e) => Logger.error(e))
+          }, config.deleteTimeout)
+        }
       })
     })
   }
